requirement_expoNet.txt
netifaces==0.11.0
speedtest-cli==2.1.3
plyer==2.0.0
